pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.121.0"
uvicorn = {extras = ["standard"], version = "^0.38.0"}
pydantic = {extras = ["email"], version = "^2.12.4"}
pydantic-settings = "^2.11.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.44"}
asyncpg = "^0.30.0"
alembic = "^1.17.1"
orjson = "^3.11.4"
jinja2 = "^3.1.6"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
itsdangerous = "^2.2.0"
python-multipart = "^0.0.20"
bcrypt = "4.0.1"
pillow = "^12.0.0"
lxml = "^6.0.2"
openpyxl = "^3.1.5"
qdrant-client = "^1.16.2"
sentence-transformers = "^5.2.0"
transformers = "^4.57.3"
accelerate = "^1.12.0"
einops = "^0.8.1"

[tool.poetry.group.dev.dependencies]
black = "^25.9.0"

